HSEARCH-2260 Made FieldHelper publicly accesible to enable further refactoring
FieldHelper is still in an impl package, though, so it should not be used outside of this maven artifact.
