+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/util/impl/FieldHelper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.impl;
+package org.hibernate.search.elasticsearch.util.impl;
 
 import java.util.Calendar;
 import java.util.Collection;
@@ -31,7 +31,7 @@ import org.hibernate.search.metadata.NumericFieldSettingsDescriptor.NumericEncod
  *
  * @author Gunnar Morling
  */
-class FieldHelper {
+public class FieldHelper {
 
 	private static final Pattern DOT = Pattern.compile( "\\." );
 
@@ -130,7 +130,7 @@ class FieldHelper {
 		}
 	}
 
-	static ExtendedFieldType getType(DocumentFieldMetadata fieldMetadata) {
+	public static ExtendedFieldType getType(DocumentFieldMetadata fieldMetadata) {
 		PropertyMetadata propertyMetata = fieldMetadata.getSourceProperty();
 		Class<?> propertyClass = propertyMetata == null ? null : propertyMetata.getPropertyClass();
 		if ( propertyClass == null ) {
@@ -186,7 +186,7 @@ class FieldHelper {
 		}
 	}
 
-	static ExtendedFieldType getType(BridgeDefinedField field) {
+	public static ExtendedFieldType getType(BridgeDefinedField field) {
 		FieldType type = field.getType();
 		if ( type == null ) {
 			return null;
@@ -211,7 +211,7 @@ class FieldHelper {
 		}
 	}
 
-	static boolean isNumeric(DocumentFieldMetadata field) {
+	public static boolean isNumeric(DocumentFieldMetadata field) {
 		if ( field.isNumeric() ) {
 			return true;
 		}
@@ -225,7 +225,7 @@ class FieldHelper {
 		return false;
 	}
 
-	static String[] getFieldNameParts(String fieldName) {
+	public static String[] getFieldNameParts(String fieldName) {
 		boolean isEmbeddedField = isEmbeddedField( fieldName );
 		return isEmbeddedField ? DOT.split( fieldName ) : new String[]{ fieldName };
 	}
